+++ NEWS.md
@@ -1,5 +1,8 @@
 # osmexport (development version)
 
+* By default, `oexp_download()` now uses a cached download if an export using 
+the same bounding box was downloaded previously.
+
 # osmexport 0.2.0
 
 * Initial release